pyproject.toml
[project]
name = "generative-prompt-engineering"
version = "0.1.0"
description = "A generative prompt engineering package for the Google Cloud Platform."
readme = "README.md"
authors = [
    { name = "kpeder", email = "kpeder@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "argparse>=1.4.0",
    "google-api-python-client>=2.178.0",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "google-cloud-logging>=3.12.1",
    "google-cloud-storage>=3.2.0",
    "google-genai>=1.29.0",
    "mcp>=1.12.4",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
]

[project.scripts]
generative-prompt-engineering = "generative_prompt_engineering:entrypoint"

[build-system]
requires = ["uv_build>=0.8.5,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "flake8>=7.3.0",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
]
